refactor: add section 관련 test code 이동
